"float *iat" should have been "int *iat".
git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@4112 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ncarg2d/man/ngmisc/ngsrat.m
@@ -8,7 +8,7 @@ CALL NGSRAT(IOPT, IAT, RAT)
 .SH C-BINDING SYNOPSIS
 #include <ncarg/ncargC.h>
 .sp
-void c_ngsrat(int iopt, float *iat, float *rat)
+void c_ngsrat(int iopt, int *iat, float *rat)
 .SH DESCRIPTION 
 .IP IOPT 12
 (an input variable of type INTEGER) indicating the desired action:
